Unserialize readdir, readdir callbacks.
A mutex is added to dir_result_t.  This lock is required
to access members.  If dirp and dirp->inode must be held
together, the inode lock is taken second.

Updates closedir, rewinddir, telldir, seekdir,
_readdir_next_frag, _readdir_rechoose_frag,
_readdir_drop_dirp_buffer, _readdir_get_frag,
_readdir_cache_cb, readdir_r_cb, _getdir_cb.

A CF_DRLOCKED flag is used to save some lock cycles, but
less aggressively than for Inode.

Signed-off-by: Matt Benjamin <matt@linuxbox.com>