Extend Connection Derivation to Match Messenger
The declaration of class Messenger as a vitual base class
nicely anticipates derivation of new message transports, but
the implementation of Connection contained artifacts (the Pipe
members) of the SimpleMessenger implementation, which is in
error.

This change solves the problem by making Connection virutal, and
pairing SimpleMessenger with a new PipeConnection, which derives
from Connection.

For performance reasons, static_cast<foo_child *>(foo *) is used
where dynamic_cast<foo_child *>(foo *) is logically intended.

Signed-off-by: Matt Benjamin <matt@linuxbox.com>