Block in xio_send_response workflow.
When a new request message sequence is decoded, the
constituent xio_msg structures are linked with the decoded
Message as a reply_hook completion.

The XioMessenger send_reply(Message *msg, Message *reply)
method replies through the reply hook.  The XioReplyHook
uses its trace of the original request to dispatch a response
sequence.

Open issues:

1. xio_client callers aren't seeing response messages delivered yet
(but the response xio_msg structures look right on entry to
xio_send_response)

2. support for response sequences larger than the request sequence
is not correct (the delta must be sent as 0..n one-way messages)

3. local delivery is not implemented--this will be prototyped as
a new local reply hook

Signed-off-by: Matt Benjamin <matt@linuxbox.com>