Fix Xio Message buffer lifetime.
Marcus found that the root cause of a buffer corruption
problem with Xio RADOS workloads was, essentially, that the
current idiom to delete buffer objects when done (and which
may be after the refcnt deletion of the corresponding Message)
led to using zero-copy buffers after recycling.

The solution chosen here is to link the lifetime of buffers to
that of the Message, such that they are disposed with the
Message completion.  This then requires that event to happen
only when all buffers have been disposed.  In addition, we would
like to pool-allocate buffer objects.  The lifetime of the
Message is shared by its XioCompletion hook, and the hook also
has an Xio memory subpool.  A new static buffer subtype is
created which holds a reference to an XioCompletion hook, and
the buffer ctor allocates the buffer itself from this pool.
Each allocated buffer counts toward the refcnt of the Message
which the hook completes, when and if it is
successfully decoded.  The new buffer destructor decrements
the linked message refcnt.  Finally the new buffers override
operator delete, which is a no-op (the buffer memory will be
reclaimed with the linked pool).

Signed-off-by: Matt Benjamin <matt@linuxbox.com>