Avoid trap on multiple discon events (same conn).
The XIO_SESSION_CONNECTION_DISCONNECTED_EVENT is being fired
twice on some/all disconnections.  Since we do non-idempotent
operations on our XioConnection handle in this case, this formerly
caused traps on the second entry.

Casey found that a workaround is to set the conn_user_ctx to NULL
the first time we see a connection disconnect, and ignore the
resulting disconnect events with event_data->conn_user_context
== NULL.

This does not fix other issues which we think may have the same
root cause (ie, the underlying cause of multiple disconnect
callbacks for the same xio_connection handle).

Signed-off-by: Matt Benjamin <matt@linuxbox.com>