Complain then discard packets that we can't possibly process.
In case of error, don't release references we don't take on xmsg.
Accelio is only giving us one notification for a chain of messages,
so there's no reason to take more than one reference on the message.
(But see immediate previous commit to src/msg/XioMessenger.cc for more on this.)

Signed-off-by: Marcus Watts <mdw@cohortfs.com>