Try-implement conditional clone of volatile buffers.
A new strong_claim method is added to buffer::list and buffer::ptr.
A strong_claim is either equivalent to claim, or else deep copies
to new raw buffer(s) if any is volatile.

A volatile buffer is one which cannot be held/claimed for long
periods due to external dependencies (e.g., Accelio message buffers
are volatile).

Signed-off-by: Matt Benjamin <matt@linuxbox.com>