Enable Accelio mempooling again.  Mark xio_msg.
The mark step allows on_msg to fix up a message that was not seen
previously in assign_data--but it's still broken because in this
case we must memdup them, I guess.  Unless Eyal changes his mind again.

Also the sequence logic is still incorrect at this commit.
The set of buffers in a logical message sequence is actually hard or
impossible to directly track in assign_data_in_buf, despite strict
ordering.  This is due to pipelining of new messages (i.e, part a new
sequence commonly) into assign_data_in_buf before a corresponding on_msg
call has been made for all messages in a given logical sequence (which
Accelio isn't even aware of, and wouldn't wish to insert barrier for,
if it were aware).  That is, the sequence boundary is not stable between
calls to assign_data_in_buf and on_msg.

Signed-off-by: Matt Benjamin <matt@linuxbox.com>