wip-addr: features: adding
I have hardwired features to be 0 here when it is encoding addresses
on the wire.  This needs to be revisited - I think async messenger
may be exchanging addresses before it knows what features the other
end supports.  If so, it will need to exchange more complete addresses
once it knows better.  But if it already has features from the
other side, it can be using that instead of 0.

Signed-off-by: Marcus Watts <mwatts@redhat.com>