http/civetweb: Don't crash if client doesn't supply host:
In theory, a client (at least that's doing HTTP/1.1) will always
set the host header field.  But that's entirely up to client software,
which we should not trust.  It's very easy to not do this using
plain telnet: "GET / HTTP/1.1"<return><return>.  The result is
a crash of the server.  That's bad.

// fixup: move this to after ld.so, before sts changes.

Signed-off-by: Marcus Watts <mwatts@redhat.com>