Fix "out of tree" builds - src/ceph.in ceph_ver.h .git_version.
When building out of tree, ceph.in is really $(srcdir)/ceph.in .
If we're building with gnu make, that can be shortened to just $<.

When building out of tree:
 - set GIT_DIR to point to source tree .git.
 - reference .git_version as $(src)/.git_version
 - never call ceph_ver.h ./ceph_ver.h

.git_version is part of "dist", so for consistency it should be in the
	source tree.
ceph_ver.h is a generated file; it belongs in the build tree.
	./ceph_ver.h is not the same thing as ceph_ver.h to make,
	and it's silly to prefix targets with ./

It would be better if configure or better yet autogen made .git_version
directly and the constructing logic just weren't in the build tree.  But
I didn't do that.

Signed-off-by: Marcus Watts <mwatts@redhat.com>