For civetweb: don't complain about trailing 's' on port number.
Remove code that complains if port is not a simple integer.

Civetweb can listen on more than one port, and port numbers can
be suffixed with 's' to indicate ssl.  There is already code elsewhere
to fetch the port string and pass it directly to civetweb.  So
the code here that parsed the port string and complained if it
wasn't a simple integer has not been doing anything useful for
a while.

Signed-off-by: Marcus Watts <mwatts@redhat.com>