fix for PEP 440 warnings when building non-release git version. XXX obselete?
Some of the extension setup.py's fetch CEPH_GIT_NICE_VER
from ceph_ver.h.  This turns out to be the result of "git describe",
and it does not meet the version string formatting requirements
of PEP 440, so generates a very scary warning message.

I've added logic to try to format the version string to be PEP
440 friendly.  So, instead of 10.0.3-1027-ge51fd74, this will make
10.0.3.post10+ge51fd74 .  This fixes a warning message at build time
for me.

The logic also carefully removes the leading 'v', since that
causes another error message, albeit less threatening.

Signed-off-by: Marcus Watts <mwatts@redhat.com>